requirements.txt
fake_useragent
fp
free_proxy
requests
selenium
selenium_wire
webdriver_manager
Django==4.2.2
django-cors-headers==4.1.0
django-decouple==2.1
django-filter==23.3
django-rest-framework==0.1.0
django-seed==0.3.1
djangorestframework==3.14.0
docutils==0.20.1
docx2pdf==0.1.8
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
whitenoise==6.6.0